+++ sample_code_submission/feature_analysis.py
@@ -1,0 +1,9 @@
+def feature_corrilations(data):
+    pass
+
+def systematics_dependence(data):
+    pass
+
+def minimal_dependent_features(data):
+    return data.columns
+
